Make t8_procidx_t long int and rename T8_MAX_CHILDREN_PER_ELE

+++ src/t8.h
@@ -99,7 +99,7 @@ typedef int32_t t8_locidx_t;
 /** Comparison function for t8_locidx_t */
 #define t8_compare_locidx(v, w) sc_int32_compare (v, w)
 /** A type for holding process ids. */
-typedef int t8_procidx_t;
+typedef int64_t t8_procidx_t;
 /** A type for global indexing that holds really big numbers. */
 typedef int64_t t8_gloidx_t;
 /** The MPI Datatype of t8_gloidx_t */
